Move elastic_stacker into its own repository
Stacker started off as a script which filled a role as part of
the Elasticsearch deployment for Livermore Computing at LC.
It had originally lived in a repo with the rest of the configuration
management system that controlled this deployment.

As part of the public release of Stacker, the Git history has been
rewritten to isolate it from the rest of the repo.