Correct httpx cert parameter in constructor
Most of the parameters from the client section of the config file are
passed verbatim into the constructor for an httpx.Client object.
However, the `cert` parameter, which is used for authentication using
X.509 certificates, doesn't fit neatly into the schema. According to
HTTPX, a cert can be either A) a path to a certificate/key encoded as
a single file, or B) a tuple of (cert, key). That's a pain to do in
Marshmallow schemas, so I had been converting from a dict to that tuple
format in a post_load hook for the schema.

However, that was inadequately tested, and when trying to merge
profiles with defaults, the schema would attempt to validate itself
twice and find the errant `cert` field in the incoming object, which had
already been validated once.

In practice, the best way around this is to fix this in the constructor so
the client can be consistent.