Hotfix: discard stats file as a separate operation
A recent change broke the transform controller's load routine. It used a set
object containing all the files to be imported, and used the `set.discard`
method to remove a metadata file from the set. However, the set.discard method
returns None and mutates the set object it was called on, rather than acting
like a set difference operation that returns the set without that element.