Remove seldom-used options
This change removes some of the options which seemed like they'd be useful
early in development, but have not been used in practice. This makes the load
and dump functions all have the same interface, which will allow us to inherit a
lot of repetitive code re: file handling, purging, etc. from the generic
controller class later.